+++ include/CoreLib/Match.hpp
@@ -121,7 +121,9 @@ namespace bw
 			const Ndk::EntityHandle& RetrieveEntityByUniqueId(EntityId uniqueId) const override;
 			EntityId RetrieveUniqueIdByEntity(const Ndk::EntityHandle& entity) const override;
 
-			void Update(float elapsedTime);
+			bool Update(float elapsedTime);
+
+			inline void Quit();
 
 			Match& operator=(const Match&) = delete;
 			Match& operator=(Match&&) = delete;
@@ -219,6 +221,7 @@ namespace bw
 			NetworkStringStore m_networkStringStore;
 			bool m_disableWhenEmpty;
 			bool m_isResetting;
+			bool m_isMatchRunning;
 	};
 }